+++ src/FaceDiscolorationManager.h
@@ -1,0 +1,12 @@
+#pragma once
+
+class FaceDiscolorationManager
+{
+public:
+	FaceDiscolorationManager() = delete;
+
+	static void InstallHooks();
+
+private:
+	static bool DataLoad_CheckRace(RE::TESNPC* a_actor);
+};
