+++ src/HeadPartsValidationManager.h
@@ -1,0 +1,9 @@
+#pragma once
+
+class HeadPartsValidationManager
+{
+public:
+	HeadPartsValidationManager() = delete;
+
+	static void InstallHooks();
+};
