+++ src/IgnorePreprocessedFaceGenManager.h
@@ -1,0 +1,9 @@
+#pragma once
+
+class IgnorePreprocessedFaceGenManager
+{
+public:
+	IgnorePreprocessedFaceGenManager() = delete;
+
+	static void InstallHooks();
+};
